refactor: moved the helper functions to a dedicated file
