+++ packages/demo/src/modules/ark-viewer/ArkViewer.tsx
@@ -7,6 +7,7 @@ import {
 	useLoaderData,
 	useRouteError,
 } from "react-router";
+import { getDocumentInfo, getEnrichment, getFulltext } from "./utils";
 
 export type ArkViewerLoaderData = {
 	document: string | null;
@@ -26,14 +27,14 @@ export const arkViewerLoader: LoaderFunction = async ({
 		};
 	}
 
-	const metadata = await getDocumentInfo(ark);
+	const documentInfo = await getDocumentInfo(ark);
 
 	const promisePool = [
-		getFulltext(metadata),
-		getEnrichment(metadata, "unitex"),
-		getEnrichment(metadata, "teeft"),
-		getEnrichment(metadata, "multicat"),
-		getEnrichment(metadata, "nb"),
+		getFulltext(documentInfo),
+		getEnrichment(documentInfo, "unitex"),
+		getEnrichment(documentInfo, "teeft"),
+		getEnrichment(documentInfo, "multicat"),
+		getEnrichment(documentInfo, "nb"),
 	];
 	const [fulltextResult, unitexResult, teeftResult, multicatResult, nbResult] =
 		await Promise.allSettled(promisePool);
@@ -68,79 +69,6 @@ export function ErrorBoundary() {
 	}
 
 	return <div>{error.message}</div>;
-}
-
-interface FileInfo {
-	mimetype: string;
-	uri: string;
-}
-
-interface DocumentInfo {
-	fulltext?: FileInfo[];
-	enrichments?: {
-		multicat?: FileInfo[];
-		nb?: FileInfo[];
-		teeft?: FileInfo[];
-		unitex?: FileInfo[];
-	};
-}
-
-interface IstexApiResponse {
-	total: number;
-	hits: DocumentInfo[];
-}
-
-async function getDocumentInfo(ark: string) {
-	const url = new URL("/document", "https://api.istex.fr");
-	url.searchParams.set("q", `arkIstex.raw:"${ark}"`);
-	url.searchParams.set("output", "fulltext,enrichments");
-
-	const response = await fetch(url);
-	if (!response.ok) {
-		throw new Error(`API responded with status ${response.status}`);
-	}
-
-	const body = (await response.json()) as IstexApiResponse;
-	if (body.total === 0 || body.hits[0] == null) {
-		throw new Error("Not found");
-	}
-
-	return body.hits[0];
-}
-
-async function getFulltext(result: DocumentInfo) {
-	const info = result.fulltext?.find(
-		(fulltext) => fulltext.mimetype === "application/tei+xml",
-	);
-	if (!info) {
-		throw new Error("No fulltext TEI");
-	}
-
-	const response = await fetch(info.uri);
-	if (!response.ok) {
-		throw new Error(`API responded with status ${response.status}`);
-	}
-
-	return await response.text();
-}
-
-async function getEnrichment(
-	result: DocumentInfo,
-	enrichmentName: keyof NonNullable<DocumentInfo["enrichments"]>,
-) {
-	const info = result.enrichments?.[enrichmentName]?.find(
-		(enrichment) => enrichment.mimetype === "application/tei+xml",
-	);
-	if (!info) {
-		throw new Error(`Enrichment ${enrichmentName} not found`);
-	}
-
-	const response = await fetch(info.uri);
-	if (!response.ok) {
-		throw new Error(`API responded with status ${response.status}`);
-	}
-
-	return await response.text();
 }
 
 function DocumentNotFound() {
